requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.1
pydantic>=2.5.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pyjwt>=2.3.0
email-validator>=2.0.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
databases[postgresql]>=0.8.0
asyncpg>=0.29.0
alembic>=1.12.1
sendgrid>=6.10.0
bcrypt>=4.0.1
